+++ index.html
@@ -10,8 +10,8 @@
 </head>
 
 <body>
+  <script src="main.js" type="module"></script>
   <main>
-
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
@@ -226,6 +226,5 @@
       </div>
     </section>
   </template>
-
 </body>
 </html>
